fix UI of currency chip

+++ tapuilibrary/src/main/java/company/tap/tapuilibrary/atoms/TapButton.kt
@@ -33,8 +33,8 @@ open class TapButton(context: Context, attributeSet: AttributeSet) :
      */
     override fun setTheme(theme: ButtonTheme) {
         theme.textColor?.let { setTextColor(it) }
-        theme.textSize?.let { textSize = it.toFloat() }
-        theme.letterSpacing?.let { letterSpacing = it.toFloat() }
+        theme.textSize?.let { textSize = it }
+        theme.letterSpacing?.let { letterSpacing = it }
         invalidate()
     }
 }
